[SOIN] Extrais la configuration du serveur
